Feat: add Nuke packaging
Package can now be done with NUKE for our Ucommerce pipelines.
This build is currently only meant for our own internal use, so OSS devs
are still recommended to use the old PowerShell scripts.

The build can update the Ucommerce nugets from filesystem with the
UpdateUcommerceNugets target. It can also download the sitecore libs
from a URL and copy them to the libs folder.

The Package target packages everything to a Sitecore package.