feature: decouple version checking from sitecore
Since our code will not necessarily be running in
in a sitecore context we need to be able to find
the sitecore version without the sitecore apis.

We have created a common interface for a sitecore
version checking service with two different
implementations, one for sitecore context and one
for cli.

sc-18474