refactor: create step for upgrading the search providers
We have taken the logic that was in the ToggleActiveSearchProviderSteps
method in PostInstallationSteps/InstallStep and moved it to a step in
the installflow.

The logic is supposed to be:

If the ElasticSerach app is enabled we need to upgrade it from the
disabled app directory, and disable the lucene app which is enabled by
default in our package.
On the other hand if ElasticSearch is NOT enabled we do nothing.

sc-18571